requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
fpdf2
bcrypt
sqlalchemy
